% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{add_to_plot}
\alias{add_to_plot}
\title{Add smoothed values to a plot}
\usage{
add_to_plot(p, .data, .key)
}
\arguments{
\item{p}{A \code{plotly} object}

\item{.data}{A data frame from a \code{forecast} object}

\item{.key}{Name of the model to add to the plot}
}
\value{
A \code{plotly} object
}
\description{
Add the smoothed time series values to a \code{plotly} plot.
}
