% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{calc_params}
\alias{calc_params}
\title{Calculate the days-on-hand}
\usage{
calc_params(.data)
}
\arguments{
\item{.data}{a data frame with daily values for inventory, shipments
received, number reprocessed, number distributed, and number returned}
}
\value{
A data frame with days-on-hand calculation.
}
\description{
Calculate the days-on-hand
}
