% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{df_to_ts}
\alias{df_to_ts}
\title{Convert a data frame to a time series}
\usage{
df_to_ts(.data, date, y, .freq = 1)
}
\arguments{
\item{.data}{data frame}

\item{date}{date column}

\item{y}{column to convert to time series}

\item{.freq}{the number of observations per unit of time}
}
\value{
Returns a \code{ts} object.
}
\description{
Converts the specified data frame column to a time series using \link[timetk]{tk_ts}.
}
