% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_mod}
\alias{get_mod}
\title{Get an individual model from a hybridModel object}
\usage{
get_mod(x, .data, .idx)
}
\arguments{
\item{x}{The name of the individual model}

\item{.data}{A \code{hybridModel} object}

\item{.idx}{A vector of dates}
}
\value{
Returns a \code{tibble} object.
}
\description{
Get an individual model from a hybridModel object
}
\keyword{internal}
