% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_value}
\alias{get_value}
\title{Get today's value}
\usage{
get_value(.data, .col, m, model = "Forecast", .date = lubridate::today())
}
\arguments{
\item{.data}{A data frame}

\item{.col}{The column with the value}

\item{m}{A string with the name of the mask}

\item{model}{A string with the name of the forecast model to use, defaults
to "Forecast" which is the ensemble model}

\item{.date}{A POSIXct date}
}
\value{
A numeric value
}
\description{
Get the value for today for a specified column
}
