% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggmod_plot}
\alias{ggmod_plot}
\title{Plot a forecast}
\usage{
ggmod_plot(.data, .y = "y", actuals = NULL)
}
\arguments{
\item{.data}{A data frame from a \code{forecast} object}

\item{.y}{A string with the column name containing the time series values.}

\item{actuals}{A data frame containing the actual (test) values.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot a forecast
}
