% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{hyb_to_df}
\alias{hyb_to_df}
\title{Converts a hybrid forecast to a data frame}
\usage{
hyb_to_df(fcast, mod, y)
}
\arguments{
\item{fcast}{An object of class \code{forecast}}

\item{mod}{An object of class \code{hybridModel}}

\item{y}{Name of the column that contains the time series values}
}
\value{
Returns a \code{tibble} object.
}
\description{
Converts a forecast object containing a ensemble (hybrid) forecast and all of
the individual forecasts, into a data frame with the ensemble forecast and
all individual forecasts.
}
