% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{make_preds}
\alias{make_preds}
\title{Make a matrix with external regressors}
\usage{
make_preds(.data, date, y, .min = NULL, .max = NULL)
}
\arguments{
\item{.data}{A data frame}

\item{date}{Name of the date column}

\item{y}{Name of the column with the external regressors}

\item{.min}{The population min value in the normalization process.}

\item{.max}{The population max value in the normalization process.}
}
\value{
Returns a matrix
}
\description{
Creates a matrix with external regressors for if the date is a weekend, and a
normalized (using \link[timetk]{normalize_vec}) vector of values to be used
in a forecasting model.
}
