% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{make_smooth}
\alias{make_smooth}
\title{Add smoothed values to data frame}
\usage{
make_smooth(.data)
}
\arguments{
\item{.data}{A data frame from a \code{forecast} object}
}
\value{
Returns a \code{tibble} object.
}
\description{
Applies the \link[stats]{loess} function to time series values and adds the
fitted values to the data frame.
}
