% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{mean_7d}
\alias{mean_7d}
\title{Calculate rolling mean}
\usage{
mean_7d(x, k = 7, fill = "extend")
}
\arguments{
\item{x}{an object (representing a series of observations)}

\item{k}{integer width of the rolling window}

\item{fill}{a three-component vector or list (recycled otherwise) providing
filling values at the left/within/to the right of the data range}
}
\value{
An object of the same class as x with the rolling mean.
}
\description{
Calculate rolling mean
}
\keyword{internal}
