% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotly_calc}
\alias{plotly_calc}
\title{Plot calculated future values}
\usage{
plotly_calc(
  .data,
  y,
  .min = NULL,
  .max = NULL,
  title = "Calculated future values",
  xtitle = "Date",
  ytitle = "Number",
  fc_line = "solid"
)
}
\arguments{
\item{.data}{A data frame containing values}

\item{y}{Column containing the values}

\item{.min}{Optional, column containing the y-min values for ribbon}

\item{.max}{Optional, column containing the y-max values for ribbon}

\item{title}{A string with the plot title.}

\item{xtitle}{A string with the x-axis title.}

\item{ytitle}{A string with the y-axis title.}

\item{fc_line}{A string which Sets the dash style of the forecast lines.
Valid options are: "solid", "dot", "dash", "longdash", "dashdot", or
"longdashdot"}
}
\value{
A \code{plotly} object
}
\description{
Plots calculated future values \code{plotly}.
}
