% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotly_fable}
\alias{plotly_fable}
\title{Plot forecasts}
\usage{
plotly_fable(
  actuals,
  y,
  combo,
  mods = NULL,
  title = "Combined forecast model",
  xtitle = "Date",
  ytitle = "Number",
  pi = TRUE,
  fc_line = "solid"
)
}
\arguments{
\item{actuals}{A tsibble object containing the known, historical, values}

\item{y}{column containing the historical values}

\item{combo}{A fable object with the forecast values}

\item{mods}{An optional fable object with values from individual forecasts}

\item{title}{A string with the plot title.}

\item{xtitle}{A string with the x-axis title.}

\item{ytitle}{A string with the y-axis title.}

\item{pi}{If \code{TRUE} the prediction intervals for all models
will be shown.}

\item{fc_line}{A string which Sets the dash style of the forecast lines.
Valid options are: "solid", "dot", "dash", "longdash", "dashdot", or
"longdashdot"}
}
\value{
A \code{plotly} object
}
\description{
Plots combined forecast model and all indivdual models using \code{plotly}.
}
