% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotly_hybrid}
\alias{plotly_hybrid}
\title{Plot a hybrid forecast}
\usage{
plotly_hybrid(
  .data,
  .title = "Combined forecast model",
  .xtitle = "Date",
  .ytitle = "Number",
  trend = TRUE
)
}
\arguments{
\item{.data}{A data frame from a \code{forecast} object, with or without smoothed values.}

\item{.title}{A string with the plot title.}

\item{.xtitle}{A string with the x-axis title.}

\item{.ytitle}{A string with the y-axis title.}

\item{trend}{If True (default), then the smoothed values are plotted; otherwise plots the actual values.}
}
\value{
A \code{plotly} object
}
\description{
Plot a hybrid forecast and all indivudal models using \code{plotly}.
}
