% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{smooth_pi}
\alias{smooth_pi}
\title{Smooth the prediction interval values}
\usage{
smooth_pi(.data, key, span = 0.75, na.action = "na.exclude")
}
\arguments{
\item{.data}{A data frame from a \code{forecast} object}

\item{key}{Column to group by}

\item{span}{Parameter which controls the degree of smoothing}

\item{na.action}{Action to be taken with missing values in the response or
predictors. The default is "na.exclude".}
}
\value{
Returns a \code{tibble} object.
}
\description{
Applies the \link[stats]{loess} function to time series values and adds the
fitted values to the data frame.
}
