% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{subset_masks}
\alias{subset_masks}
\title{Create a subset of mask data}
\usage{
subset_masks(.data, masks, ...)
}
\arguments{
\item{.data}{data frame}

\item{masks}{vector of masks}

\item{...}{grouping columns}
}
\value{

}
\description{
Create a subset of mask data
}
