% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{ts_to_df}
\alias{ts_to_df}
\title{Convert a time series to a data frame}
\usage{
ts_to_df(.ts, timetk_idx = TRUE, dates = NULL)
}
\arguments{
\item{.ts}{A time-series forecast of class \code{forecast}}

\item{timetk_idx}{If timetk index (non-regularized index) is present, uses it
to develop forecast. Otherwise uses default index.}

\item{dates}{vector of dates to replace non-POSIXct indices}
}
\value{
Returns a \code{tibble} object.
}
\description{
Converts a time series back into a date frame using \link[sweep]{sw_sweep}.
}
