% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{inventory_calcs}
\alias{inventory_calcs}
\title{Calculate inventory-related values}
\usage{
inventory_calcs(.data, .xreg, .col)
}
\arguments{
\item{.data}{A data frame with historical values}

\item{.xreg}{A data frame with forecast values}

\item{.col}{Column with the number of masks distributed}
}
\value{
Returns a \code{tibble} object.
}
\description{
Calculate the expected inventory, burn rate, days remaining, and
days-on-hand.
}
