% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotly_smooth}
\alias{plotly_smooth}
\title{Plot a smoothed forecast}
\usage{
plotly_smooth(
  .data,
  title = "Combined forecast model",
  xtitle = "Date",
  ytitle = "Number",
  width = 1200,
  height = 400
)
}
\arguments{
\item{.data}{A data frame containing the historical and forecast values}

\item{title}{A string with the plot title.}

\item{xtitle}{A string with the x-axis title.}

\item{ytitle}{A string with the y-axis title.}

\item{width}{Width in pixels}

\item{height}{Height in pixels}
}
\value{
A \code{plotly} object
}
\description{
Plot a forecast with smoothed trend line using \code{plotly}.
}
